pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = "A contacts management API using FastAPI"
authors = ["Lada Murchych <lada11lada@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pydantic = "^2.10.3"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
email-validator = "^2.2.0"
faker = "^33.1.0"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
python-jose = "^3.3.0"
fastapi-mail = "^1.4.2"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21.0"
mypy = "^1.0.0"
flake8 = "^6.0.0"
black = "^23.0.0"
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
